:adhesive_bandage: (sonarcloud): Fix implicit conversion loses integer precision
https://sonarcloud.io/project/issues?id=leka_LekaOS&issues=AXiE6yzYQJJ2tBYqGH4h&open=AXiE6yzYQJJ2tBYqGH4h&pullRequest=140

+++ drivers/LKCoreVideo/include/internal/CGPixel.h
@@ -14,8 +14,8 @@
 namespace leka {
 
 struct Point {
-	uint16_t x = 0;
-	uint16_t y = 0;
+	uint32_t x = 0;
+	uint32_t y = 0;
 };
 
 struct CGPixel {
